Fix errors, improve notation, better results
This commit fixes numerous errors. In particular it fixes an erroneous result for separable states. It also improves notation and explanation. There are now good complete explanations for all of the special cases discussed.

The exception is mixed, non-separable states with k = 2. For this case, the current explanation is correct and cohesive, but further progress should be made to derive nicer results.